feat: Upgrade mongo image and include env variable PORT in the docker-compose file
Upgrade the image mongo to stable tag 4.4.10. The version is compatible to lib mongoose.
I removed the bind to port in the mongo service, because I think that is unnecessary to
expose port to the external host.

I include the env variable PORT to facilite change the node service port via .env environment variable